terrario-serra-backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.116.1"}
psycopg = {extras = ["binary"], version = "^3.2.9"}
sqlalchemy = "^2.0.43"
alembic = "^1.16.4"
pydantic-settings = "^2.10.1"
python-dotenv = "^1.1.1"
apscheduler = "^3.11.0"
httpx = "^0.28.1"
tinytuya = "^1.17.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
